Rename sklearn.py
Resolves #18
