+++ pytsetlini/__init__.py
@@ -10,4 +10,4 @@ __version__ = '0.0.1'
 # add any imports here, if you wish to bring things into the library's
 # top-level namespace when the library is imported.
 
-from .sklearn import TsetlinMachineClassifier
+from .sklearn_estimator import TsetlinMachineClassifier
